tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-delay_inline.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-delay and inline element</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-delay-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-delay applies to inline elements">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes fade {
         from {
            opacity: 1;
         }
         to {
            opacity: 0;
         }
      }
      span {
         color: blue;
      }
      .test {
         animation-duration: 0.5s;
         animation-delay: 500ms;
         animation-fill-mode: forwards;
         animation-name: fade;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var span = document.querySelector('span');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(span, null);
         var passed = false;
         function aboutEqual(observed, actual) {
            var maxDeviation = 100;
            return (Math.abs(observed - actual) < maxDeviation);
         }
         var timeout = setTimeout(function() {
            test(function(){assert_true(false);}, testName); done();
            log.data = 'FAIL';
         }, 1500);
         setTimeout(function() {
            passed = (style.getPropertyValue('opacity') == '1');
         }, 200);
         span.addEventListener('animationend', function() {
            clearTimeout(timeout);
            passed = (passed && style.getPropertyValue('opacity') == '0');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         });
         span.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue word below should start fading out after half a second and should
   be completely invisible after another half a second..</p>
   <p>Testing...</p>
   <div>normal <span>FADE</span> normal</div>
</body>
</html>
